Deprecate epd_powerdown() and make it to a board specific function instead
