+++ src/epd_driver/display_ops.c
@@ -40,15 +40,12 @@ void epd_poweron() {
   epd_board->poweron(&ctrl_state);
 }
 
-#if defined(CONFIG_EPD_BOARD_REVISION_LILYGO_T5_47)
-void epd_powerdown() {
-  cfg_powerdown(&config_reg);
-  i2s_gpio_detach();
-}
-#endif
-
 void epd_poweroff() {
   epd_board->poweroff(&ctrl_state);
+}
+
+epd_ctrl_state_t *epd_ctrl_state() {
+  return &ctrl_state;
 }
 
 void epd_deinit() {
